requirements.txt
torch
diffusers
transformers
scipy
Pillow
matplotlib
opencv-python
